rename variables and files
